requirements.txt
elasticsearch 
pyyaml